Factor out common components of TryBlock and CatchBlock
