+++ src/Block/TryCatch/CatchBlock.php
@@ -5,12 +5,10 @@ declare(strict_types=1);
 namespace webignition\BasilCompilableSource\Block\TryCatch;
 
 use webignition\BasilCompilableSource\Body\BodyInterface;
-use webignition\BasilCompilableSource\HasMetadataInterface;
 use webignition\BasilCompilableSource\Line\CatchExpression;
-use webignition\BasilCompilableSource\Metadata\Metadata;
 use webignition\BasilCompilableSource\Metadata\MetadataInterface;
 
-class CatchBlock implements HasMetadataInterface
+class CatchBlock extends AbstractBlock
 {
     private const RENDER_TEMPLATE = <<<'EOD'
 catch (%s) {
@@ -19,48 +17,31 @@ catch (%s) {
 EOD;
 
     private CatchExpression $catchExpression;
-    private BodyInterface $body;
-
 
     public function __construct(CatchExpression $catchExpression, BodyInterface $body)
     {
+        parent::__construct($body);
+
         $this->catchExpression = $catchExpression;
-        $this->body = $body;
+    }
+
+    protected function getRenderTemplate(): string
+    {
+        return self::RENDER_TEMPLATE;
+    }
+
+    protected function getAdditionalRenderComponents(): array
+    {
+        return [
+            $this->catchExpression->render(),
+        ];
     }
 
     public function getMetadata(): MetadataInterface
     {
-        $metadata = new Metadata();
+        $metadata = parent::getMetadata();
         $metadata = $metadata->merge($this->catchExpression->getMetadata());
-        $metadata = $metadata->merge($this->body->getMetadata());
 
         return $metadata;
-    }
-
-    public function render(): string
-    {
-        $renderedBody = $this->body->render();
-
-        $renderedBody = $this->indent($renderedBody);
-        $renderedBody = rtrim($renderedBody, "\n");
-
-        return sprintf(self::RENDER_TEMPLATE, $this->catchExpression->render(), $renderedBody);
-    }
-
-    private function indent(string $content): string
-    {
-        if ('' === $content) {
-            return '';
-        }
-
-        $lines = explode("\n", $content);
-
-        array_walk($lines, function (&$line) {
-            if ('' !== $line) {
-                $line = '    ' . $line;
-            }
-        });
-
-        return implode("\n", $lines);
     }
 }
